requirements.txt
opencv-python
numpy
scipy
matplotlib
streamlit
Pillow